requirements.txt
genanki
python-dotenv
httpx
openai
pydantic
pyYAML